% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{baixar_datajud}
\alias{baixar_datajud}
\title{Pega uma base zip do Datajud}
\usage{
baixar_datajud(tribunal, oj, path)
}
\arguments{
\item{tribunal}{codigo do tribunal.}

\item{oj}{código da unidade judiciária.}

\item{path}{caminho da pasta para salvar o arquivo zip}
}
\value{
caminho do arquivo.
}
\description{
Pega uma base zip do Datajud
}
