% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{listar_municipios}
\alias{listar_municipios}
\title{Listar os municípios}
\usage{
listar_municipios(id_tribunal, path)
}
\arguments{
\item{id_tribunal}{id do tribunal para baixar os municipios.}

\item{path}{pasta para salvar o arquivo json.}
}
\value{
caminho para um arquivo do json que contém os municípios.
}
\description{
Listar os municípios
}
