% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{listar_oj}
\alias{listar_oj}
\title{Listar Órgãos de Justiça}
\usage{
listar_oj(id_tribunal, id_municipio, path)
}
\arguments{
\item{id_tribunal}{id do tribunal para baixar os municípios}

\item{id_municipio}{id do municípiio para baixar ojs}

\item{path}{pasta para salvar o arquivo json.}
}
\value{
caminho do arquivo json que contém ojs.
}
\description{
Listar Órgãos de Justiça
}
