% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{listar_tribunais}
\alias{listar_tribunais}
\title{Listar tribunais a partir da API do DataJud}
\usage{
listar_tribunais()
}
\value{
tibble com a lista dos tribunais.
}
\description{
Listar tribunais a partir da API do DataJud
}
