% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{municipios}
\alias{municipios}
\title{municipios}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5586 rows and 3 columns.
}
\usage{
municipios
}
\description{
municipios
}
\keyword{datasets}
