% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{ojs}
\alias{ojs}
\title{ojs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20836 rows and 12 columns.
}
\usage{
ojs
}
\description{
ojs
}
\keyword{datasets}
