% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{tribunais}
\alias{tribunais}
\title{tribunais}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 91 rows and 3 columns.
}
\usage{
tribunais
}
\description{
tribunais
}
\keyword{datasets}
